fix: satisfy DOGFOOD localization and review gates
