+++ examples/docs/counter-block-demo.part03.ts
@@ -1,5 +1,6 @@
 import { boxSurface, graphemeWidth, type BijouContext, type Surface } from '@flyingrobots/bijou';
 import { contentSurface } from '../_shared/example-surfaces.js';
+import { dogfoodLocalizedText as dogfoodText } from './localization.js';
 import { normalizeCounterMode, normalizeCounterValue } from './counter-block-demo.part01.js';
 import type { CounterDemoBlockConfig, CounterDemoModel } from './counter-block-demo.part01.js';
 import { counterDemoBlock } from './counter-block-demo.part02.js';
@@ -8,6 +9,7 @@ export function counterDemoBlockConfig(
   model: CounterDemoModel,
   ctx: BijouContext,
   width: number,
+  localization?: CounterDemoBlockConfig['localization'],
 ): CounterDemoBlockConfig {
   return {
     counter: model.counter,
@@ -15,6 +17,7 @@ export function counterDemoBlockConfig(
     animationTimeMs: model.animationTimeMs,
     width,
     ctx,
+    localization,
   };
 }
 
@@ -27,7 +30,17 @@ export function counterDemoBlockSurface(config: CounterDemoBlockConfig): Surface
   const output = counterDemoBlock.render({ config, mode }).output;
   const lines = output.split('\n');
   return boxSurface(contentSurface(output), {
-    title: mode === 'static' ? 'CounterDemoBlock static' : 'CounterDemoBlock fixture',
+    title: mode === 'static'
+      ? dogfoodText(
+        config.localization,
+        'counterDemo.surfaceTitle.static',
+        'CounterDemoBlock static',
+      )
+      : dogfoodText(
+        config.localization,
+        'counterDemo.surfaceTitle.fixture',
+        'CounterDemoBlock fixture',
+      ),
     width: Math.max(32, Math.min(72, config.width ?? widestLine(lines) + 4)),
     borderToken: config.ctx?.border('primary'),
     padding: { left: 1, right: 1 },
