+++ examples/docs/i18n/source/dogfood-strings.csv
@@ -627,6 +627,14 @@ bijou.dogfood,blocks.surfaceInventory.title,message,en,string,DOGFOOD Surface Bl
 bijou.dogfood,blocks.surfaceInventory.title,message,en,string,DOGFOOD Surface Blocks,en,string,DOGFOOD Surface Blocks,current,
 bijou.dogfood,blocks.surfaceInventory.title,message,en,string,DOGFOOD Surface Blocks,es,,,missing,
 bijou.dogfood,blocks.surfaceInventory.title,message,en,string,DOGFOOD Surface Blocks,fr,,,missing,
+bijou.dogfood,counterDemo.surfaceTitle.fixture,message,en,string,CounterDemoBlock fixture,de,string,CounterDemoBlock-Test-Fixture,current,
+bijou.dogfood,counterDemo.surfaceTitle.fixture,message,en,string,CounterDemoBlock fixture,en,string,CounterDemoBlock fixture,current,
+bijou.dogfood,counterDemo.surfaceTitle.fixture,message,en,string,CounterDemoBlock fixture,es,string,fixture de CounterDemoBlock,current,
+bijou.dogfood,counterDemo.surfaceTitle.fixture,message,en,string,CounterDemoBlock fixture,fr,string,fixture CounterDemoBlock,current,
+bijou.dogfood,counterDemo.surfaceTitle.static,message,en,string,CounterDemoBlock static,de,string,CounterDemoBlock statisch,current,
+bijou.dogfood,counterDemo.surfaceTitle.static,message,en,string,CounterDemoBlock static,en,string,CounterDemoBlock static,current,
+bijou.dogfood,counterDemo.surfaceTitle.static,message,en,string,CounterDemoBlock static,es,string,CounterDemoBlock estático,current,
+bijou.dogfood,counterDemo.surfaceTitle.static,message,en,string,CounterDemoBlock static,fr,string,CounterDemoBlock statique,current,
 bijou.dogfood,docs.empty.coverage.body,message,en,string,DOGFOOD currently documents {documented} of {total} canonical component families. This field guide is honest about current coverage and will keep expanding over time.,de,,,missing,
 bijou.dogfood,docs.empty.coverage.body,message,en,string,DOGFOOD currently documents {documented} of {total} canonical component families. This field guide is honest about current coverage and will keep expanding over time.,en,string,DOGFOOD currently documents {documented} of {total} canonical component families. This field guide is honest about current coverage and will keep expanding over time.,current,
 bijou.dogfood,docs.empty.coverage.body,message,en,string,DOGFOOD currently documents {documented} of {total} canonical component families. This field guide is honest about current coverage and will keep expanding over time.,es,,,missing,
@@ -1527,6 +1535,10 @@ bijou.dogfood,stories.preview.viewport.width,message,en,string,width={value},de,
 bijou.dogfood,stories.preview.viewport.width,message,en,string,width={value},en,string,width={value},current,
 bijou.dogfood,stories.preview.viewport.width,message,en,string,width={value},es,string,ancho={value},current,
 bijou.dogfood,stories.preview.viewport.width,message,en,string,width={value},fr,string,largeur={value},current,
+bijou.dogfood,storybook.catalog.title,message,en,string,catalog,de,string,Katalog,current,
+bijou.dogfood,storybook.catalog.title,message,en,string,catalog,en,string,catalog,current,
+bijou.dogfood,storybook.catalog.title,message,en,string,catalog,es,string,catálogo,current,
+bijou.dogfood,storybook.catalog.title,message,en,string,catalog,fr,string,catalogue,current,
 bijou.dogfood,terminal.nonInteractive.requiresTty,message,en,string,DOGFOOD requires an interactive terminal with both stdin and stdout attached to a TTY.,de,string,DOGFOOD benötigt ein interaktives Terminal mit stdin und stdout an einem TTY.,current,
 bijou.dogfood,terminal.nonInteractive.requiresTty,message,en,string,DOGFOOD requires an interactive terminal with both stdin and stdout attached to a TTY.,en,string,DOGFOOD requires an interactive terminal with both stdin and stdout attached to a TTY.,current,
 bijou.dogfood,terminal.nonInteractive.requiresTty,message,en,string,DOGFOOD requires an interactive terminal with both stdin and stdout attached to a TTY.,es,string,DOGFOOD requiere una terminal interactiva con stdin y stdout conectados a un TTY.,current,
