+++ packages/bijou-mcp/src/tools/docs.part04.ts
@@ -1,10 +1,52 @@
 import { z } from 'zod';
 import { buildStructuredToolResult, structuredToolOutputShape, withOutputMode } from '../output.js';
-import type { ToolRegistration } from '../types.js';
-import { MCP_DOCS_CATALOG } from './docs-catalog/index.js';
+import type { ToolRegistration, ToolResult } from '../types.js';
+import { DEFAULT_DOCS_ONLY_INTERACTION_PROFILES, DEFAULT_INTERACTION_PROFILES, MCP_DOCS_CATALOG } from './docs-catalog/index.js';
+import type { ToolDocsCatalogEntry, ToolInteractionProfiles } from './docs-catalog/index.js';
 import type { SerializedToolDocsEntry } from './docs.part01.js';
 import { DOCS_ONLY_EXAMPLE_RENDERERS, normalizeDocsTerm } from './docs.part02.js';
-import { exampleText, resolvedInteractionProfiles, scoreDocsEntry } from './docs.part03.js';
+
+function scoreDocsEntry(entry: ToolDocsCatalogEntry, normalizedQuery: string): number {
+  if (normalizedQuery === '') return 1;
+  const primaryFields = [
+    entry.toolName,
+    entry.family,
+    entry.category,
+    ...entry.aliases,
+  ].map(normalizeDocsTerm);
+  const secondaryFields = [
+    entry.summary,
+    ...entry.related,
+  ].map(normalizeDocsTerm);
+  if (primaryFields.some(value => value === normalizedQuery)) return 100;
+  if (primaryFields.some(value => value.startsWith(normalizedQuery))) return 80;
+  let score = primaryFields.some(value => value.includes(normalizedQuery)) ? 40 : 0;
+  if (secondaryFields.some(value => value === normalizedQuery)) score += 20;
+  if (secondaryFields.some(value => value.includes(normalizedQuery))) score += 10;
+  const tokens = normalizedQuery.split(/\s+/).filter(Boolean);
+  for (const token of tokens) {
+    if (primaryFields.some(value => value.includes(token))) score += 10;
+    else if (secondaryFields.some(value => value.includes(token))) score += 4;
+  }
+  return score;
+}
+
+function exampleText(result: ToolResult): string {
+  return result.content[0]?.text
+    ?? (typeof result.structuredContent?.['rendered'] === 'string'
+      ? result.structuredContent['rendered']
+      : '');
+}
+
+function resolvedInteractionProfiles(
+  entry: ToolDocsCatalogEntry,
+  mcpExposed: boolean,
+): ToolInteractionProfiles {
+  return {
+    ...(mcpExposed ? DEFAULT_INTERACTION_PROFILES : DEFAULT_DOCS_ONLY_INTERACTION_PROFILES),
+    ...entry.interactionProfiles,
+  };
+}
 
 export function createDocsTool(tools: readonly ToolRegistration[]): ToolRegistration {
   const toolMap = new Map(tools.map(tool => [tool.name, tool]));
