+++ packages/bijou-node/src/recorder.ts
@@ -1,3 +1,2 @@
 export type { NativeDemoSpec, RecorderResult, SurfaceGifOptions } from './recorder.part01.js';
-export { rasterizeSurface, writeSurfaceGif } from './recorder.part03.js';
-export { recordDemoGif } from './recorder.part04.js';
+export { rasterizeSurface, recordDemoGif, writeSurfaceGif } from './recorder.part03.js';
