+++ packages/bijou-tui/src/canvas.ts
@@ -1,2 +1,2 @@
 export type { CanvasOptions, CanvasResolution, ShaderCell, ShaderFn, ShaderParams } from './canvas.part01.js';
-export { canvas } from './canvas.part04.js';
+export { canvas } from './canvas.part03.js';
