+++ packages/create-bijou-tui-app/src/index.part01.ts
@@ -109,4 +109,27 @@ function toPackageName(input: string): string {
   return normalized.length > 0 ? normalized : DEFAULT_TARGET_DIR;
 }
 
+/**
+ * Normalize target directory and derive package name.
+ */
+export function resolveTarget(targetDir: string | undefined, cwd: string): {
+  readonly targetDir: string;
+  readonly absTargetDir: string;
+  readonly packageName: string;
+} {
+  const normalizedTarget = targetDir?.trim().length
+    ? targetDir
+    : DEFAULT_TARGET_DIR;
+
+  const absTargetDir = resolve(cwd, normalizedTarget);
+  const base = basename(absTargetDir) || DEFAULT_TARGET_DIR;
+
+  return {
+    targetDir: normalizedTarget,
+    absTargetDir,
+    packageName: toPackageName(base),
+  };
+}
+
 export { toPackageName };
+import { basename, resolve } from 'node:path';
