+++ packages/create-bijou-tui-app/src/index.ts
@@ -1,5 +1,4 @@
 export type { PackageManager, ParsedArgs, ScaffoldOptions, ScaffoldResult } from './index.part01.js';
-export { DEFAULT_TARGET_DIR, detectPackageManager, parseArgs } from './index.part01.js';
-export { resolveTarget } from './index.part02.js';
+export { DEFAULT_TARGET_DIR, detectPackageManager, parseArgs, resolveTarget } from './index.part01.js';
 export { createTemplateFiles } from './index.part03.js';
 export { scaffoldProject, usage } from './index.part04.js';
